+++ node/lib/internal/bootstrap_node.js
@@ -48,9 +48,7 @@
       NativeModule.require('internal/process/write-coverage').setup();
 
     if (process.env.ENCLOSE_IO_USE_ORIGINAL_NODE) {
-      if (!process.env.ENCLOSE_IO_ALWAYS_USE_ORIGINAL_NODE) {
-        delete process.env.ENCLOSE_IO_USE_ORIGINAL_NODE;
-      }
+      delete process.env.ENCLOSE_IO_USE_ORIGINAL_NODE;
     }
 
     // Do not initialize channel in debugger agent, it deletes env variable